+++ src/main/java/com/alexis/vitraux/item/BlueprintItem.java
@@ -1,0 +1,21 @@
+package com.alexis.vitraux.item;
+
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.item.tooltip.TooltipType;
+import net.minecraft.text.Text;
+
+import java.util.List;
+
+public class BlueprintItem extends Item {
+
+    public BlueprintItem(Settings settings) {
+        super(settings);
+    }
+
+    @Override
+    public void appendTooltip(ItemStack stack, Item.TooltipContext context, List<Text> tooltip, TooltipType type) {
+        super.appendTooltip(stack, context, tooltip, type);
+        TemplateItem.appendCanvasTooltip(stack, tooltip);
+    }
+}